+++ index.js
@@ -2,6 +2,7 @@
 
 module.exports = {
 	extends: ['stylelint-config-standard-scss', 'stylelint-config-suitcss'],
+	plugins: ['stylelint-high-performance-animation'],
 	rules: {
 		// redeclared from standard-scss because suitcss sets to `always`
 		'block-closing-brace-newline-after': [
@@ -35,6 +36,7 @@ module.exports = {
 				ignorePattern: '/https?://[0-9,a-z]*.*/',
 			},
 		],
+		'plugin/no-low-performance-animation-properties': [true, { ignore: 'paint-properties' }],
 		'rule-empty-line-before': [
 			'always-multi-line',
 			{